+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Unreleased
 
+## 1.39.0 (April 20, 2021)
+
+BUG FIXES:
+
+* resource/spotinst_ocean_aks: retry failed import
+
 ## 1.38.0 (April 12, 2021)
 
 FEATURES: